# docs/swagger.yaml
basePath: /v1
definitions:
  models.Event:
    properties:
      date:
        type: string
      deleted:
        type: boolean
      flyer:
        type: string
      id:
        type: integer
      location:
        type: string
      name:
        type: string
    required:
    - date
    - name
    type: object
  models.Video:
    properties:
      category:
        $ref: '#/definitions/models.VideoCategory'
      id:
        type: integer
      subtitle:
        type: string
      title:
        type: string
      videoSrc:
        type: string
    required:
    - category
    - title
    - videoSrc
    type: object
  models.VideoCategory:
    enum:
    - ORIGINAL SONG
    - COVER
    type: string
    x-enum-varnames:
    - CategoryOriginalSong
    - CategoryCover
host: localhost:5000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para gerenciamento de eventos e shows da banda Retrovisionarios.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Retrovisionarios API
  version: "1.0"
paths:
  /events:
    get:
      description: Retorna uma lista de eventos, opcionalmente filtrada por ano e
        status de deleção.
      parameters:
      - description: Ano para filtrar eventos
        in: query
        name: year
        type: integer
      - description: Se true, lista também eventos deletados
        in: query
        name: deleted
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.Event'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar eventos
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Cria um novo evento de show.
      parameters:
      - description: Dados do evento
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.Event'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Event'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar evento
      tags:
      - events
  /events/{id}:
    delete:
      description: Realiza um soft delete de um evento.
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deletar evento
      tags:
      - events
    get:
      description: Retorna os detalhes de um único evento baseado no ID informado.
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obter evento por ID
      tags:
      - events
    patch:
      consumes:
      - application/json
      description: Atualiza as propriedades de um evento existente.
      parameters:
      - description: ID do evento
        in: path
        name: id
        required: true
        type: integer
      - description: Dados do evento para atualizar
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/models.Event'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar evento
      tags:
      - events
  /videos:
    get:
      description: Retorna uma lista de videos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.Video'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar videos
      tags:
      - videos
swagger: "2.0"